+++ core-ng/src/main/java/core/framework/module/WebSocketConfig.java
@@ -1,11 +1,11 @@
 package core.framework.module;
 
 import core.framework.internal.inject.InjectValidator;
-import core.framework.internal.module.Config;
 import core.framework.internal.module.ModuleContext;
 import core.framework.internal.web.HTTPIOHandler;
 import core.framework.internal.web.websocket.ChannelHandler;
 import core.framework.internal.web.websocket.WebSocketHandler;
+import core.framework.internal.web.websocket.WebSocketMetrics;
 import core.framework.web.websocket.ChannelListener;
 import core.framework.web.websocket.WebSocketContext;
 import org.slf4j.Logger;
@@ -16,14 +16,12 @@ import java.time.Duration;
 /**
  * @author neo
  */
-public final class WebSocketConfig extends Config {
+public final class WebSocketConfig {
+    final ModuleContext context;
     private final Logger logger = LoggerFactory.getLogger(WebSocketConfig.class);
-    ModuleContext context;
 
-    @Override
-    protected void initialize(ModuleContext context, String name) {
+    public WebSocketConfig(ModuleContext context) {
         this.context = context;
-        context.httpServer.handler.rateControl.config(WebSocketHandler.WS_OPEN_GROUP, 10, 10, Duration.ofSeconds(30));
     }
 
     public <T, V> void listen(String path, Class<T> clientMessageClass, Class<V> serverMessageClass, ChannelListener<T, V> listener) {
@@ -40,6 +38,9 @@ public final class WebSocketConfig extends Config {
         if (context.httpServer.handler.webSocketHandler == null) {
             context.httpServer.handler.webSocketHandler = new WebSocketHandler(context.logManager, context.httpServer.siteManager.sessionManager, context.httpServer.handler.rateControl);
             context.beanFactory.bind(WebSocketContext.class, null, context.httpServer.handler.webSocketHandler.context);
+
+            context.httpServer.handler.rateControl.config(WebSocketHandler.WS_OPEN_GROUP, 10, 10, Duration.ofSeconds(30));
+            context.collector.metrics.add(new WebSocketMetrics(context.httpServer.handler.webSocketHandler));
         }
 
         context.beanClassValidator.validate(clientMessageClass);