+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### 8.0.13 (02/08/2023 - )
 
+* ws: add "ws_active_channels" stats
+
 ### 8.0.12 (01/18/2023 - 02/07/2023)
 
 * hash: removed hmac md5/sha1, hash.sha1 support